+++ lib/src/tile_source.dart
@@ -2,11 +2,15 @@ import 'dart:ui';
 
 import 'themes/sprite.dart';
 import 'tileset.dart';
+import 'tileset_raster.dart';
 
 /// a model representing the source data of a tile.
 class TileSource {
   /// the tileset of the tile
   final Tileset tileset;
+
+  /// the raster tileset of the tile, for raster layers of the theme.
+  final RasterTileset rasterTileset;
 
   /// The optional sprite index, which provides sprites referenced by the theme.
   /// If absent, sprites are ignored when rendering.
@@ -16,5 +20,9 @@ class TileSource {
   /// If absent, sprites are ignored when rendering.
   final Image? spriteAtlas;
 
-  TileSource({required this.tileset, this.spriteIndex, this.spriteAtlas});
+  TileSource(
+      {required this.tileset,
+      this.spriteIndex,
+      this.spriteAtlas,
+      this.rasterTileset = const RasterTileset(tiles: {})});
 }